+++ lib/src/posthog.dart
@@ -1,6 +1,7 @@
 import 'package:meta/meta.dart';
 
 import 'package:posthog_flutter/src/error_tracking/posthog_error_tracking_autocapture_integration.dart';
+import 'package:posthog_flutter/src/error_tracking/posthog_exception.dart';
 import 'posthog_config.dart';
 import 'posthog_flutter_platform_interface.dart';
 import 'posthog_observer.dart';
@@ -83,6 +84,8 @@ class Posthog {
   Future<void> capture({
     required String eventName,
     Map<String, Object>? properties,
+    Map<String, Object>? userProperties,
+    Map<String, Object>? userPropertiesSetOnce,
   }) {
     final propertiesCopy = properties == null ? null : {...properties};
 
@@ -95,6 +98,8 @@ class Posthog {
     return _posthog.capture(
       eventName: eventName,
       properties: propertiesCopy,
+      userProperties: userProperties,
+      userPropertiesSetOnce: userPropertiesSetOnce,
     );
   }
 
@@ -172,6 +177,22 @@ class Posthog {
           Map<String, Object>? properties}) =>
       _posthog.captureException(
           error: error, stackTrace: stackTrace, properties: properties);
+
+  /// Captures runZonedGuarded exceptions with optional custom properties
+  /// https://api.flutter.dev/flutter/dart-async/runZonedGuarded.html
+  ///
+  /// [error] - The error/exception to capture
+  /// [stackTrace] - Optional stack trace (if not provided, current stack trace will be used)
+  /// [properties] - Optional custom properties to attach to the exception event
+  Future<void> captureRunZonedGuardedError(
+      {required Object error,
+      StackTrace? stackTrace,
+      Map<String, Object>? properties}) async {
+    final wrappedError = PostHogException(
+        source: error, mechanism: 'runZonedGuarded', handled: false);
+    await _posthog.captureException(
+        error: wrappedError, stackTrace: stackTrace, properties: properties);
+  }
 
   /// Closes the PostHog SDK and cleans up resources.
   ///